Flesh out check-out-for-transmit process
Create stored function check_out_enqueued_for_transmit() to encapsulate
the tricky-ish writable CTE with a SELECT ... FOR UPDATE first portion.

The logic there is to find N unsent records of the current desired
vector and then lock them w/o any race conditions. Use SELECT ...
FOR UPDATE to be sure to find unlocked and not-about-to-be-locked
records (if we've somehow got multiple cronjobs running simultaneously)
as feeder for an update statement which then visibly marks them
as locked, ending with returing that set of records to the caller.

Documented how this is intented to be used over in the schema
file, and added the additional procedural control columns after
having fully remembered all of the nuances.